svg/painting/parsing/color-interpolation-computed.svg

<svg xmlns="http://www.w3.org/2000/svg" xmlns:h="http://www.w3.org/1999/xhtml" width="800" height="800"><title>SVG Painting: getComputedStyle().colorInterpolation</title><metadata><h:link href="https://svgwg.org/svg2-draft/painting.html#ColorInterpolationProperty" rel="help"/><h:meta content="color-interpolation computed value is as specified." name="assert"/></metadata><h:script src="/resources/testharness.js"/><h:script src="/resources/testharnessreport.js"/><h:script src="/css/support/computed-testcommon.js"/><script><![CDATA[

test_computed_value("color-interpolation", "auto");
test_computed_value("color-interpolation", "srgb");
test_computed_value("color-interpolation", "linearrgb");

  ]]></script></svg>
