LayoutTests/imported/w3c/web-platform-tests/svg/painting/parsing/color-interpolation-computed-expected.txt


PASS Property color-interpolation value 'auto'
FAIL Property color-interpolation value 'srgb' assert_equals: expected "srgb" but got "sRGB"
FAIL Property color-interpolation value 'linearrgb' assert_equals: expected "linearrgb" but got "linearRGB"
